pyproject.toml
[project]
name = "rebot-8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
    "google-genai>=1.26.0",
    "ruff>=0.12.4",
]
